% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Checkout Git repositories}
\usage{
checkout(repos)
}
\arguments{
\item{repos}{Path to one or more Git repositories.}
}
\value{
Called for its side effect. Returns \code{repos} invisibly.
}
\description{
Checkout the \code{main} or \code{master} branch of the given Git repositories. But
stay on the current reference (e.g. branch) if checking out the repository
that called \code{checkout()}.
}
