% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkoutInput.R
\name{checkoutInput}
\alias{checkoutInput}
\title{Stripe Checkout Shiny Input}
\usage{
checkoutInput(id, amount = 100, name = "Stripe.com", description = "",
  publickey = NULL, label = "Pay with Card",
  image = paste0("https://stripe.com/img/documentation",
  "/checkout/marketplace.png"), button_id = "customButton")
}
\arguments{
\item{id}{Shiny id for the output token}

\item{amount}{Amount to be charged}

\item{name}{Name of the organization on the checkout app}

\item{description}{Description on the checkout app}

\item{publickey}{Stripe API key}

\item{label}{Label on the Checkout Button}

\item{image}{Image on the checkout app}

\item{button_id}{JS ID for the button used}
}
\value{
The result from these functions is a tag object.
}
\description{
Form wrapper for Stripe Checkout
}
