% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler_script.R
\name{handler_script}
\alias{handler_script}
\title{Stripe Checkout Handler Script}
\usage{
handler_script(amount, name, description, button_id)
}
\arguments{
\item{amount}{Amount to be charged}

\item{name}{Name of the organization on the checkout app}

\item{description}{Description on the checkout app}

\item{button_id}{JS ID for the button used}
}
\value{
The result is a character of JavaScript
}
\description{
Wraps JS for the handler script for Stripe
}
