% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_script.R
\name{token_script}
\alias{token_script}
\title{Stripe Checkout Token Script}
\usage{
token_script(id, publickey, image)
}
\arguments{
\item{id}{Shiny id for the output token}

\item{publickey}{Stripe API key}

\item{image}{Image on the checkout app}
}
\value{
The result is a character of JavaScript
}
\description{
Wraps JS for the token script for Stripe
}
