% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Checkout Git repositories}
\usage{
checkout(repos)
}
\arguments{
\item{repos}{Path to a Git repository}
}
\value{
Called for its side effect. Returns \code{repos} invisibly.
}
\description{
Checkout the \code{HEAD} of the current Git repository.
}
