% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkout.R
\name{checkout}
\alias{checkout}
\title{Checkout Git repositories}
\usage{
checkout(repos)
}
\arguments{
\item{repos}{Path to one or more Git repositories.}
}
\value{
Called for its side effect. Returns \code{repos} invisibly.
}
\description{
Checkout the \code{main} or \code{master} branch of the given Git repositories. But
stay on the current reference (e.g. branch) if checking out the repository
that called \code{checkout()}.
}
\examples{
library(magrittr)

# Helper
walk <- function(x, f, ...) {
  lapply(x, f, ...)
  invisible(x)
}

# Setup two minimal repositories.
repos <- file.path(tempdir(), paste0("repo", 1:2))
repos \%>\% walk(dir.create)

repos \%>\%
  git("init --initial-branch=main") \%>\%
  git("add .") \%>\%
  git("config user.name Jerry") \%>\%
  git("config user.email jerry@gmail.com") \%>\%
  git("commit -m 'Initialize' --allow-empty") \%>\%
  git("log --oneline -n 1", verbose = TRUE)

# If we set the directory at `repo1`, it stays at the branch `pr`, whereas the
# `repo2` changes to the branch `master` (or `main`).
withr::with_dir(repos[[1]], {
  repos \%>\% git("checkout -b pr")
  # Compare before and after `checkout()`
  # Before checkout(), both repos are at the branch pr
  repos \%>\% git("branch", verbose = TRUE)
  repos \%>\% checkout()
  # After checkout(), the repo1 is at the branch pr and repo2 is at main
  repos \%>\% git("branch", verbose = TRUE)
})

walk(repos, unlink, recursive = TRUE)
}
